move methods out of management command code
- move is_domain_deleted into Domain class as staticmethod
- move ucr db methods into dbaccessors
